ui/pyproject.toml
[project]
name = "UI"
version = "0.2.0"
description = "Web UI for Chat Gourmet AI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai==1.44.1",
    "streamlit==1.38.0",
    "psycopg2-binary>=2.9.9",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.8",
]
